fix(pricing): read edition/trajectory price as cents (canonical)
The edition/trajectory price field stores CENTS (admin enters euros, save
×100, display ÷100 — a self-consistent cents system). But the public read
path treated the stored value as euros (Money::eur), so €4,95 (stored 495
cents) rendered as €495,00. Canonical unit = CENTS (matches the Money VO's
internal representation + the admin round-trip).

Read sites fixed euros→cents (no-site-left-behind):
- EditionService::getPrice — Money::eur → Money::cents (THE main bug)
- EnrollmentFormHandler (trajectory quote + Money::eur)
- QuoteAdminController (dropped a ×100 double-conversion — found via ground-truth)
- card-edition.php (catalog edition card also read cents as euros — found via ground-truth)
- card-trajectory.php, trajectory-card.php, single-vad_trajectory.php, enrollment.php
Sites already cents-correct (÷100 / getPrice→Money / raw-cents) left untouched.

Seeder now stores cents (×100 at edition + trajectory + seed-quote store points),
so future seeds match canonical.

Migration script scripts/migrations/2026-06-30-price-euros-to-cents.php
(dry-run-safe, APPLY=1 to run) converts the _stride_seed_data rows
(seeded euros) ×100; the 61 admin-written rows are already cents and untouched.

Tier A money logic, RED-first: getPrice(stored 495) now returns € 4,95 (was
€ 495,00). Unit 1150→1151, 0 fails; PHPStan + Pint clean.

+++ web/app/themes/stridence/helpers/trajectory-card.php
@@ -45,7 +45,7 @@ if (!function_exists('stridence_build_trajectory_card_args')) {
             'status'         => $trajectory['status'] ?? 'open',
             'course_count'   => $service->getCourseCount($trajectoryId),
             'elective_count' => $service->getElectiveGroupCount($trajectoryId),
-            'price'          => (float) ($trajectory['price'] ?? 0),
+            'price'          => (int) ($trajectory['price'] ?? 0), // canonical cents
             'deadline'       => (string) ($trajectory['enrollment_deadline'] ?? ''),
             // Enrolled-only (dashboard) — null/'' on the catalog path.
             'progress'       => $progress,
